+++ src/Reducers/signUp.js
