+++ src/utils/colors.js
@@ -1,0 +1,17 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+
+//Added by Aran Bacall 25/09/2017 CGI Colours
+var cgiRed = exports.cgiRed = 'rgba(227,25,55,1)';
+var cgiBeet = exports.cgiBeet = '#991f3d';
+var cgiPumpkin = exports.cgiPumpkin = '#ff6a00';
+var cgiHoney = exports.cgiHoney = '#f2a200';
+var cgiCloud = exports.cgiCloud = '#a5acb0';
+var cgiIce = exports.cgiIce = '#a1c4d0';
+
+
+
+
